requirements.txt
pandas
numpy
joblib
scikit-learn
xgboost
shap
fastapi
uvicorn
pydantic
pytest
httpx
azure-identity
azure-keyvault-secrets
azure-storage-blob
